Move common controllers into `common.controller`
