+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2/controller/Cas2ReportsController.kt
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2v2ReportName
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.service.Cas2ReportsService
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.controller.generateXlsxStreamingResponse
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.common.controller.generateXlsxStreamingResponse
 
 @Cas2Controller
 class Cas2ReportsController(private val cas2v2ReportService: Cas2ReportsService) {
