+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas3/controller/Cas3ReportsController.kt
@@ -22,12 +22,12 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.reporting.propertie
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.reporting.properties.FutureBookingsReportProperties
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.reporting.properties.TransitionalAccommodationReferralReportProperties
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.service.Cas3ReportService
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.common.controller.ContentType
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.common.controller.generateStreamingResponse
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.common.controller.generateXlsxStreamingResponse
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.common.problem.BadRequestProblem
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.common.problem.ForbiddenProblem
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.common.problem.ParamDetails
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.controller.ContentType
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.controller.generateStreamingResponse
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.controller.generateXlsxStreamingResponse
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.FeatureFlagService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.UserAccessService
 import java.time.LocalDate
