+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/common/controller/ControllerUtils.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.controller
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.common.controller
 
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
