+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/common/controller/PeopleController.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.controller
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.common.controller
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
