+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/common/controller/TasksController.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.controller
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.common.controller
 
 import jakarta.transaction.Transactional
 import org.springframework.http.ResponseEntity
